+++ app/models/message.js
@@ -20,7 +20,25 @@ export default class Message extends BaseModel {
       "created_at",
 
       "status", // virtual
+      "reactions", // virtual
     ]
+  }
+
+  static get allowedAttachmentFields() {
+    return [
+      "file_id",
+      "file_name",
+      "file_url",
+      "file_blur_hash",
+      "file_content_type",
+      "file_size",
+      "file_width",
+      "file_height",
+    ]
+  }
+
+  static get allowedBotAttachmentFields() {
+    return ["file_url", "file_blur_hash", "file_content_type", "file_size", "file_width", "file_height"]
   }
 
   static get hiddenFields() {