+++ app/providers/services/message/index.js
@@ -1,9 +1,12 @@
+import { ERROR_STATUES } from "../../../constants/errors.js"
+
 class MessageService {
-  constructor(helpers, userRepo, messageRepo, messageStatusRepo) {
+  constructor(helpers, userRepo, messageRepo, messageStatusRepo, messageReactionRepo) {
     this.helpers = helpers
     this.userRepo = userRepo
     this.messageRepo = messageRepo
     this.messageStatusRepo = messageStatusRepo
+    this.messageReactionRepo = messageReactionRepo
   }
 
   async create(user, conversation, blockedUserIds, messageParams) {
@@ -19,25 +22,49 @@ class MessageService {
     return message
   }
 
-  async processHandlerResult(accept, baseMessage, message, options) {
+  #validateAttachmentInMessage(attachment) {
+    const allowedAttachmentFields = this.messageRepo.Model.allowedBotAttachmentFields
+
+    if (typeof attachment !== "object" || attachment === null) return false
+    return Object.keys(attachment).every((key) => allowedAttachmentFields.includes(key))
+  }
+
+  async processHandlerResult(organizationId, accept, baseMessage, message, options) {
     if (accept) return {}
 
-    const { body } = message
-    if (body && typeof body === "string") {
-      if (options.isReplaceBody) {
-        const newMessageFields = { body }
-        return { newMessageFields }
-      } else {
-        const existServerBot = await this.userRepo.findByLogin("server-chat-bot")
-        if (existServerBot) {
-          const botMessageParams = { ...baseMessage, body }
-          return { botMessageParams, serverBot: existServerBot }
+    const { body, attachments } = message
+    const processedResponse = {}
+
+    if (attachments?.length) {
+      for (const att of attachments) {
+        if (!this.#validateAttachmentInMessage(att)) {
+          throw new Error(ERROR_STATUES.INVALID_ATTACHMENT_FIELDS.message, {
+            cause: ERROR_STATUES.INVALID_ATTACHMENT_FIELDS,
+          })
         }
       }
-    } else {
-      //error that body is must be string
     }
-    return {}
+
+    if (body && typeof body !== "string") {
+      throw new Error(ERROR_STATUES.INCORRECT_TYPE_OF_BODY.message, {
+        cause: ERROR_STATUES.INCORRECT_TYPE_OF_BODY,
+      })
+    }
+
+    if (!body?.length && !attachments?.length) return processedResponse
+
+    if (options.isReplaceBody) {
+      processedResponse.newMessageFields = { body, attachments }
+      return processedResponse
+    }
+
+    const existServerBot = await this.userRepo.findByLogin(organizationId, "server-chat-bot")
+    if (existServerBot) {
+      processedResponse.botMessageParams = { ...baseMessage, body, attachments }
+      processedResponse.serverBot = existServerBot
+    }
+
+    return processedResponse
   }
 
   async messagesList(cId, user, options, limit) {
@@ -55,7 +82,9 @@ class MessageService {
 
     const messagesStatuses = await this.messageStatusRepo.findReadStatusForMids(messageIds)
 
-    return { messages, messagesStatuses }
+    const messagesReactions = await this.messageReactionRepo.aggregateForUserMessages(messageIds, user.native_id)
+
+    return { messages, messagesStatuses, messagesReactions }
   }
 
   async hasAccessToMessage(messageId, userId) {
@@ -111,11 +140,15 @@ class MessageService {
 
     const lastMessagesIds = Object.values(aggregateLastMessage).map((msg) => msg._id)
 
-    const aggregateLastMessageStatus = await this.messageStatusRepo.findReadStatusForMids(lastMessagesIds)
+    const messagesStatuses = await this.messageStatusRepo.findReadStatusForMids(lastMessagesIds)
+    const messagesReactions = await this.messageReactionRepo.aggregateForUserMessages(lastMessagesIds, user?.native_id)
 
     Object.values(aggregateLastMessage).forEach((message) => {
-      const status = aggregateLastMessageStatus[message._id.toString()] ? "read" : "sent"
+      const status = messagesStatuses[message._id] ? "read" : "sent"
       message.set("status", status)
+
+      const reactions = messagesReactions[message._id] ?? {}
+      message.set("reactions", reactions)
     })
 
     return aggregateLastMessage
@@ -134,6 +167,21 @@ class MessageService {
     )
 
     return unreadMessageCountByCids
+  }
+
+  async updateReactions(mid, userId, addReaction, removeReaction) {
+    const updated = { add: false, remove: false }
+
+    if (addReaction) {
+      const upsert = { mid, user_id: userId, reaction: addReaction }
+      updated.add ||= await this.messageReactionRepo.upsert(upsert)
+    }
+
+    if (removeReaction) {
+      updated.remove ||= await this.messageReactionRepo.remove(mid, userId, removeReaction)
+    }
+
+    return updated
   }
 
   async deleteMessages(userId, mIds, deleteAll) {